Login: Move login code to firebaseauthactivity
