+++ app/src/main/java/ua/safetynet/auth/FirebaseAuthActivity.java
@@ -3,10 +3,14 @@ package ua.safetynet.auth;
 import android.content.Intent;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
+import android.util.Log;
 import android.widget.Toast;
 
 import com.firebase.ui.auth.AuthUI;
+import com.firebase.ui.auth.ErrorCodes;
 import com.firebase.ui.auth.IdpResponse;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 
 
 import java.util.Arrays;
@@ -21,49 +25,61 @@ import ua.safetynet.user.MainPageActivity;
 
 public class FirebaseAuthActivity extends AppCompatActivity
 {
-
-    private static final int RC_SIGN_IN = 100;
+    FirebaseAuth firebaseAuth;
+    FirebaseUser firebaseUser;
+    private static final int RCSIGNIN = 123;
 
     @Override
     protected void onCreate(Bundle savedInstanceState)
     {
         super.onCreate(savedInstanceState);
 
-        List<AuthUI.IdpConfig> providers = Arrays.asList(
-                new AuthUI.IdpConfig.EmailBuilder().build(),
-                new AuthUI.IdpConfig.PhoneBuilder().build(),
-                new AuthUI.IdpConfig.GoogleBuilder().build());
+        firebaseAuth = FirebaseAuth.getInstance();
+        firebaseUser = firebaseAuth.getCurrentUser();
 
-        startActivityForResult(
-                AuthUI.getInstance()
-                        .createSignInIntentBuilder()
-                        .setAvailableProviders(providers)
-                        .build(),
-                RC_SIGN_IN);
+        if (firebaseUser == null) //If the user isnt already logged in
+        {
+            //User isnt signed in, launch sign in activity
+            startActivityForResult(AuthUI.getInstance()
+                    .createSignInIntentBuilder()
+                    .setAvailableProviders(Arrays.asList(
+                            new AuthUI.IdpConfig.GoogleBuilder().build(),
+                            new AuthUI.IdpConfig.EmailBuilder().build(),
+                            new AuthUI.IdpConfig.PhoneBuilder().build()))
+                    .setIsSmartLockEnabled(false)
+                    .build(),RCSIGNIN);
+        }
     }
 
 
     // [START auth_fui_result]
     @Override
-    protected void onActivityResult(int requestCode, int resultCode, Intent data)
-    {
+    public void onActivityResult(int requestCode, int resultCode, Intent data) {
         super.onActivityResult(requestCode, resultCode, data);
+        // RC_SIGN_IN is the request code you passed into startActivityForResult(...) when starting the sign in flow.
+        if (requestCode == RCSIGNIN) {
+            IdpResponse response = IdpResponse.fromResultIntent(data);
 
-        if (resultCode == RESULT_OK)
-        {
             // Successfully signed in
-            Intent in = new Intent(this, MainPageActivity.class);
-            in.putExtra("EXTRA_IDP_RESPONSE", IdpResponse.fromResultIntent(data));
-            startActivity(in);
-            finish();
-            return;
-        }
-        if(resultCode == RESULT_CANCELED)
-        {
-            Toast.makeText(getApplicationContext(),R.string.login_user_cancelled,Toast.LENGTH_LONG).show();
-            return;
-        }
+            if (resultCode == RESULT_OK) {
+                startActivity(new Intent(this,MainPageActivity.class));
+                finish();
+            } else {
+                // Sign in failed
+                if (response == null) {
+                    // User pressed back button
+                    startActivity(new Intent(this, SplashScreenActivity.class));
+                    return;
+                }
 
+                if (response.getError().getErrorCode() == ErrorCodes.NO_NETWORK) {
+                    Toast.makeText(this, "Network Required to login", Toast.LENGTH_SHORT).show();
+                    startActivity(new Intent(this, SplashScreenActivity.class));
+                    return;
+                }
+                Log.e("Splash Screen Login", "Sign-in error: ", response.getError());
+            }
+        }
     }
 }
