+++ app/src/main/java/ua/safetynet/auth/SplashScreenActivity.java
@@ -20,70 +20,16 @@ import ua.safetynet.user.MainPageActivity;
 
 public class SplashScreenActivity extends AppCompatActivity
 {
-    FirebaseAuth firebaseAuth;
-    FirebaseUser firebaseUser;
-    private static final int RCSIGNIN = 123;
+
     @Override
     protected void onCreate(Bundle savedInstanceState)
     {
         super.onCreate(savedInstanceState);
-        firebaseAuth = FirebaseAuth.getInstance();
-        firebaseUser = firebaseAuth.getCurrentUser();
 
+        Intent intent = new Intent(this, MainPageActivity.class);
+        startActivity(intent);
+        finish();
 
-        if (firebaseUser == null) //If the user isnt already logged in
-        {
-            //User isnt signed in, launch sign in activity
-            startActivityForResult(AuthUI.getInstance()
-                    .createSignInIntentBuilder()
-                    .setAvailableProviders(Arrays.asList(
-                            new AuthUI.IdpConfig.GoogleBuilder().build(),
-                            new AuthUI.IdpConfig.EmailBuilder().build(),
-                            new AuthUI.IdpConfig.PhoneBuilder().build()))
-                    .setIsSmartLockEnabled(false)
-                    .build(),RCSIGNIN);
-        }
-        //Register auth state listener to start this activity over again on user sign out so they will have to sign in again
-        /*firebaseAuth.addAuthStateListener(new FirebaseAuth.AuthStateListener() {
-            @Override
-            public void onAuthStateChanged(@NonNull FirebaseAuth firebaseAuth) {
-                if(firebaseAuth.getCurrentUser() == null) {
-                    startActivity(new Intent(getBaseContext(), SplashScreenActivity.class));
-                }
-            }
-        });*/
-        else {
-            Intent intent = new Intent(this, MainPageActivity.class);
-            startActivity(intent);
-            finish();
-        }
-    }
-
-    @Override
-    public void onActivityResult(int requestCode, int resultCode, Intent data) {
-        super.onActivityResult(requestCode, resultCode, data);
-        // RC_SIGN_IN is the request code you passed into startActivityForResult(...) when starting the sign in flow.
-        if (requestCode == RCSIGNIN) {
-            IdpResponse response = IdpResponse.fromResultIntent(data);
-
-            // Successfully signed in
-            if (resultCode == RESULT_OK) {
-                startActivity(new Intent(this,MainPageActivity.class));
-                finish();
-            } else {
-                // Sign in failed
-                if (response == null) {
-                    // User pressed back button
-                    return;
-                }
-
-                if (response.getError().getErrorCode() == ErrorCodes.NO_NETWORK) {
-                    Toast.makeText(this, "Network Required to login", Toast.LENGTH_SHORT).show();
-                    return;
-                }
-                Log.e("Splash Screen Login", "Sign-in error: ", response.getError());
-            }
-        }
     }
 }
